feat: set up webpack build configuration, remove unused config files
