+++ package.json
@@ -4,10 +4,9 @@
   "description": "JavaScript SDK for Veriff identity verification",
   "main": "lib/veriff.js",
   "scripts": {
-    "build": "webpack --config webpack/prod.config.js",
-    "dev": "webpack --progress --colors --watch --config webpack/prod.config.js",
-    "test": "./node_modules/karma/bin/karma start ./scripts/karma.conf.js",
-    "ts-build": "webpack --config webpack/webpack.config.js --mode production"
+    "build": "webpack --config webpack/webpack.config.js --mode production",
+    "dev": "webpack --progress --colors --watch --config webpack/webpack.config.js",
+    "test": "./node_modules/karma/bin/karma start ./scripts/karma.conf.js"
   },
   "author": "Ilia Aphtsiauri <ilia@veriff.me> (https://github.com/iaptsiauri) && Maksim Tsvetkov <maksim.tsvetkov@veriff.com> (https://github.com/MaxFlower)",
   "keywords": [
@@ -45,7 +44,7 @@
     "typescript": "^3.7.2",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10",
-    "webpack-merge": "^4.1.3",
+    "webpack-merge": "^4.2.2",
     "yargs": "^11.1.0"
   },
   "dependencies": {}
