+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "outDir": "./dist",
     "allowJs": true,
+    "module": "commonjs",
     "target": "es5",
     "lib": ["es2015", "es2017", "dom"],
     "declaration": true
