+++ webpack/webpack.config.js
@@ -1,12 +1,20 @@
-module.exports = {
-    entry: "./src/index.ts",
-    output: {
-        filename: "./lib/veriff.js",
-        libraryTarget: "umd"
-    },
+const libraryName = 'Veriff';
+const merge = require('webpack-merge');
 
+const baseConfig = {
+    entry: './src/index.ts',
     resolve: {
-        extensions: [".ts", ".tsx", ".js", ".css"]
+        extensions: ['.ts', '.js', '.css']
+    },
+};
+
+const umdConfig = merge(baseConfig, {
+    output: {
+        path: `${__dirname}/../dist`,
+        filename: 'veriff.min.js',
+        library: libraryName,
+        libraryTarget: 'umd',
+        umdNamedDefine: true
     },
 
     module: {
@@ -21,7 +29,30 @@ module.exports = {
                 use: ['style-loader', 'css-loader'],
             },
         ],
+    }
+});
+
+const npmConfig = merge(baseConfig, {
+    output: {
+        path: `${__dirname}/../lib`,
+        filename: 'veriff.js',
+        library: libraryName,
+        libraryTarget: 'commonjs2'
     },
 
-    // Other options...
-};
+    module: {
+        rules: [
+            {
+                test: /\.tsx?$/,
+                use: 'awesome-typescript-loader',
+                exclude: /node_modules/,
+            },
+            {
+                test: /\.css$/,
+                use: ['style-loader', 'css-loader'],
+            },
+        ],
+    }
+});
+
+module.exports = [umdConfig, npmConfig];
